+++ package.json
@@ -39,16 +39,16 @@
   "contributes": {
     "commands": [
       {
-        "command": "expo.config.open.prebuild.file",
+        "command": "expo.config.prebuild.preview",
         "title": "Expo: Preview modifier"
       },
       {
-        "command": "expo.config.open.prebuild.file.json",
+        "command": "expo.config.prebuild.preview.json",
         "title": "Expo: Preview modifier as JSON"
       },
       {
-        "command": "expo.config.open.prebuild.config",
-        "title": "Expo: Preview prebuild config"
+        "command": "expo.config.preview",
+        "title": "Expo: Preview config"
       }
     ]
   },
